requirements.txt
requests>=2.31.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pyngrok>=6.0.0 